server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.7
alembic==1.12.1
pydantic==2.12.5
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
fastapi-mail==1.4.1
jinja2==3.1.2
passlib[sha256_crypt]==1.7.4
python-jose[cryptography]==3.3.0
akshare>=1.12.0
pandas>=2.0.0
numpy>=1.24.0
apscheduler>=3.10.0
